requirements.txt
fastapi
uvicorn[standard]
google-genai
numpy
